Authentication system added and nutrition details screen implemented

+++ app/src/main/java/com/kumela/cmeter/model/nutrition/AltMeasure.java
@@ -15,7 +15,7 @@ public class AltMeasure {
 
     public String measure;
     public int seq;
-    public int qty;
+    public float qty;
 
     @NonNull
     @Override
